Добавляет модульность
